Update README.md
Without the import path it was kind of confusing reading the code. I was using a previous version that did not need the package so i was surprised when i looked at the code and didn't see where the variable `jwt` was coming from, until i checked the example file.